requirements.txt
aiosmtpd==1.4.6
alembic==1.16.1
atpublic==6.0.1
attrs==25.3.0
autopep8==2.3.2
babel==2.17.0
black==25.1.0
blinker==1.9.0
certifi==2025.7.9
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.0
dnspython==2.7.0
elastic-transport==8.17.1
elasticsearch==9.0.2
email_validator==2.2.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
googletrans==4.0.0rc1
greenlet==3.2.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
polib==1.2.0
pycodestyle==2.13.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
rfc3986==1.5.0
ruff==0.11.11
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
